requirements.txt
astroid==2.14.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dill==0.3.6
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
pylint==2.16.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
